// src/types/imageConfig.ts
export interface ImageConfig {
    data: Uint8Array;            
    opacity?: number;       
    showShadow?: boolean;    
    shadowColor?: string;    
    shadowBlur?: number;     
    shadowOffsetX?: number;  
    shadowOffsetY?: number;  
    shadowSpread?: number;   
    showBorder?: boolean;
    borderColor?: string;
    borderWidth?: number;
    borderRadius?: number;  
    width?: number;         
    height?: number;      
    duration?: number;       
  }